# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Caramurú Field Engineering API",
    "description": "API da plataforma de engenharia de campo da Caramurú Construções. Multi-tenant, obras (subestações e LTs) com PostGIS e ativos com QR Code.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/sync": {
      "get": {
        "tags": [
          "sync"
        ],
        "summary": "Pull Changes",
        "description": "Consistent tenant delta. V1 supports schema 1, without schema migrations.",
        "operationId": "pull_changes_api_v1_sync_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "last_pulled_at",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9007199254740991,
              "minimum": 0,
              "default": 0,
              "title": "Last Pulled At"
            }
          },
          {
            "name": "schema_version",
            "in": "query",
            "required": false,
            "schema": {
              "const": 1,
              "type": "integer",
              "default": 1,
              "title": "Schema Version"
            }
          },
          {
            "name": "migration",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "const": "null",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Migration"
            }
          },
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sync"
        ],
        "summary": "Push Changes",
        "description": "Atomically apply a durable mutation batch; replay the same batch_id after network failure.",
        "operationId": "push_changes_api_v1_sync_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Tenant-ID",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JsonValue": {},
      "MergeConflict": {
        "properties": {
          "table_name": {
            "type": "string",
            "enum": [
              "projects",
              "assets",
              "inspections",
              "aprs"
            ],
            "title": "Table Name"
          },
          "record_id": {
            "type": "string",
            "format": "uuid",
            "title": "Record Id"
          },
          "rejected_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Rejected Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "table_name",
          "record_id",
          "rejected_fields"
        ],
        "title": "MergeConflict"
      },
      "MutationMetadata": {
        "properties": {
          "client_version": {
            "type": "string",
            "maxLength": 85,
            "pattern": "^[0-9]{13}:[0-9]{6}:[a-zA-Z0-9_-]{1,64}$",
            "title": "Client Version"
          },
          "changed_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Changed Fields"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "client_version"
        ],
        "title": "MutationMetadata"
      },
...
      "PullResponse": {
        "properties": {
          "changes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/PullTableChanges"
            },
            "propertyNames": {
              "enum": [
                "projects",
                "assets",
                "inspections",
                "aprs"
              ]
            },
            "type": "object",
            "title": "Changes"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "changes",
          "timestamp"
        ],
        "title": "PullResponse"
      },
      "PullTableChanges": {
        "properties": {
          "created": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Created"
          },
          "updated": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Updated"
          },
          "deleted": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deleted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PullTableChanges"
      },
      "PushRequest": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "last_pulled_at": {
            "type": "integer",
            "maximum": 9007199254740991.0,
            "minimum": 0.0,
            "title": "Last Pulled At"
          },
          "changes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TableChanges"
            },
            "propertyNames": {
              "enum": [
                "projects",
                "assets",
                "inspections",
                "aprs"
              ]
            },
            "type": "object",
            "title": "Changes"
          },
          "metadata": {
            "additionalProperties": {
              "additionalProperties": {
                "$ref": "#/components/schemas/MutationMetadata"
              },
              "propertyNames": {
                "format": "uuid"
              },
              "type": "object"
            },
            "propertyNames": {
              "enum": [
                "projects",
                "assets",
                "inspections",
                "aprs"
              ]
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "last_pulled_at",
          "changes",
          "metadata"
        ],
        "title": "PushRequest"
      },
      "PushResponse": {
        "properties": {
          "batch_id": {
            "type": "string",
            "format": "uuid",
            "title": "Batch Id"
          },
          "applied": {
            "type": "integer",
            "title": "Applied"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/MergeConflict"
            },
            "type": "array",
            "title": "Conflicts"
          },
          "requires_pull": {
            "type": "boolean",
            "title": "Requires Pull",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "batch_id",
          "applied",
          "conflicts"
        ],
        "title": "PushResponse"
      },
...
      "TableChanges": {
        "properties": {
          "created": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Created"
          },
          "updated": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Updated"
          },
          "deleted": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Deleted"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TableChanges"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}